requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
Pillow==10.0.1
pandas==2.1.1
reportlab==4.0.4
scikit-learn==1.3.1
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
celery>=5.3.6
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0